backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
pytest
pytest-asyncio
httpx
email-validator
